chore(ui): extract labelled props types and create utilities
