--- packages/react-ui/src/classnames.ts
@@ -1,13 +1,0 @@
-import classNames, { Argument } from 'classnames';
-
-export { classNames };
-
-/**
- * The props types for classnames.
- */
-export type ClassNames = {
-  /**
-   * The element className.
-   */
-  className?: Argument;
-};
